Struct rusoto_logs::CloudWatchLogsClient

pub struct CloudWatchLogsClient { /* fields omitted */ }

A client for the Amazon CloudWatch Logs API.

Methods

impl CloudWatchLogsClient

pub fn new(region: Region) -> CloudWatchLogsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CloudWatchLogsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,