Struct rusoto_logs::FilterLogEventsRequest

pub struct FilterLogEventsRequest {
    pub end_time: Option<i64>,
    pub filter_pattern: Option<String>,
    pub interleaved: Option<bool>,
    pub limit: Option<i64>,
    pub log_group_name: String,
    pub log_stream_names: Option<Vec<String>>,
    pub next_token: Option<String>,
    pub start_time: Option<i64>,
}

Fields
end_time: Option<i64>

The end of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not returned.

filter_pattern: Option<String>

The filter pattern to use. If not provided, all the events are matched.

interleaved: Option<bool>

If the value is true, the operation makes a best effort to provide responses that contain events from multiple log streams within the log group interleaved in a single response. If the value is false all the matched log events in the first log stream are searched first, then those in the next log stream, and so on. The default is false.

limit: Option<i64>

The maximum number of events to return. The default is 10,000 events.

log_group_name: String

The name of the log group.

log_stream_names: Option<Vec<String>>

Optional list of log stream names.

next_token: Option<String>

The token for the next set of events to return. (You received this token from a previous call.)

start_time: Option<i64>

The start of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp prior to this time are not returned.