Crate rusoto_logs

Amazon CloudWatch Logs

If you're using the service, you're probably looking for CloudWatchLogsClient and CloudWatchLogs.

Structs

CancelExportTaskRequest

CloudWatchLogsClient
A client for the Amazon CloudWatch Logs API.

CreateExportTaskRequest

CreateExportTaskResponse

CreateLogGroupRequest

CreateLogStreamRequest

DeleteDestinationRequest

DeleteLogGroupRequest

DeleteLogStreamRequest

DeleteMetricFilterRequest

DeleteRetentionPolicyRequest

DeleteSubscriptionFilterRequest

DescribeDestinationsRequest

DescribeDestinationsResponse

DescribeExportTasksRequest

DescribeExportTasksResponse

DescribeLogGroupsRequest

DescribeLogGroupsResponse

DescribeLogStreamsRequest

DescribeLogStreamsResponse

DescribeMetricFiltersRequest

DescribeMetricFiltersResponse

DescribeSubscriptionFiltersRequest

DescribeSubscriptionFiltersResponse

Destination
Represents a cross-account destination that receives subscription log events.

ExportTask
Represents an export task.

ExportTaskExecutionInfo
Represents the status of an export task.

ExportTaskStatus
Represents the status of an export task.

FilterLogEventsRequest

FilterLogEventsResponse

FilteredLogEvent
Represents a matched event.

GetLogEventsRequest

GetLogEventsResponse

InputLogEvent
Represents a log event, which is a record of activity that was recorded by the application or resource being monitored.

ListTagsLogGroupRequest

ListTagsLogGroupResponse

LogGroup
Represents a log group.

LogStream
Represents a log stream, which is a sequence of log events from a single emitter of logs.

MetricFilter
Metric filters express how CloudWatch Logs would extract metric observations from ingested log events and transform them into metric data in a CloudWatch metric.

MetricFilterMatchRecord
Represents a matched event.

MetricTransformation
Indicates how to transform ingested log events into metric data in a CloudWatch metric.

OutputLogEvent
Represents a log event.

PutDestinationPolicyRequest

PutDestinationRequest

PutDestinationResponse

PutLogEventsRequest

PutLogEventsResponse

PutMetricFilterRequest

PutRetentionPolicyRequest

PutSubscriptionFilterRequest

RejectedLogEventsInfo
Represents the rejected events.

SearchedLogStream
Represents the search status of a log stream.

SubscriptionFilter
Represents a subscription filter.

TagLogGroupRequest

TestMetricFilterRequest

TestMetricFilterResponse

UntagLogGroupRequest

Enums

CancelExportTaskError
Errors returned by CancelExportTask

CreateExportTaskError
Errors returned by CreateExportTask

CreateLogGroupError
Errors returned by CreateLogGroup

CreateLogStreamError
Errors returned by CreateLogStream

DeleteDestinationError
Errors returned by DeleteDestination

DeleteLogGroupError
Errors returned by DeleteLogGroup

DeleteLogStreamError
Errors returned by DeleteLogStream

DeleteMetricFilterError
Errors returned by DeleteMetricFilter

DeleteRetentionPolicyError
Errors returned by DeleteRetentionPolicy

DeleteSubscriptionFilterError
Errors returned by DeleteSubscriptionFilter

DescribeDestinationsError
Errors returned by DescribeDestinations

DescribeExportTasksError
Errors returned by DescribeExportTasks

DescribeLogGroupsError
Errors returned by DescribeLogGroups

DescribeLogStreamsError
Errors returned by DescribeLogStreams

DescribeMetricFiltersError
Errors returned by DescribeMetricFilters

DescribeSubscriptionFiltersError
Errors returned by DescribeSubscriptionFilters

FilterLogEventsError
Errors returned by FilterLogEvents

GetLogEventsError
Errors returned by GetLogEvents

ListTagsLogGroupError
Errors returned by ListTagsLogGroup

PutDestinationError
Errors returned by PutDestination

PutDestinationPolicyError
Errors returned by PutDestinationPolicy

PutLogEventsError
Errors returned by PutLogEvents

PutMetricFilterError
Errors returned by PutMetricFilter

PutRetentionPolicyError
Errors returned by PutRetentionPolicy

PutSubscriptionFilterError
Errors returned by PutSubscriptionFilter

TagLogGroupError
Errors returned by TagLogGroup

TestMetricFilterError
Errors returned by TestMetricFilter

UntagLogGroupError
Errors returned by UntagLogGroup

Traits

CloudWatchLogs
Trait representing the capabilities of the Amazon CloudWatch Logs API. Amazon CloudWatch Logs clients implement this trait.

Type Definitions

AccessPolicy

Arn

Days
The number of days to retain the log events in the specified log group. Possible values are: 1, 3, 5, 7, 14, 30, 60, 90, 120, 150, 180, 365, 400, 545, 731, 1827, and 3653.

DefaultValue

Descending

DescribeLimit

DestinationArn

DestinationName

Destinations

Distribution

EventId

EventMessage

EventNumber

EventsLimit

ExportDestinationBucket

ExportDestinationPrefix

ExportTaskId

ExportTaskName

ExportTaskStatusCode

ExportTaskStatusMessage

ExportTasks

ExtractedValues

FilterCount

FilterName

FilterPattern
A symbolic description of how CloudWatch Logs should interpret the data in each log event. For example, a log event may contain timestamps, IP addresses, strings, and so on. You use the filter pattern to specify what to look for in the log event message.

FilteredLogEvents

InputLogEvents

InputLogStreamNames

Interleaved

LogEventIndex

LogGroupName

LogGroups

LogStreamName

LogStreamSearchedCompletely

LogStreams

MetricFilterMatches

MetricFilters

MetricName
The name of the CloudWatch metric to which the monitored log information should be published. For example, you may publish to a metric called ErrorCount.

MetricNamespace

MetricTransformations

MetricValue
The value to publish to the CloudWatch metric. For example, if you're counting the occurrences of a term like "Error", the value is "1" for each occurrence. If you're counting the bytes transferred, the value is the value in the log event.

NextToken
The token for the next set of items to return. The token expires after 24 hours.

OrderBy

OutputLogEvents

RoleArn

SearchedLogStreams

SequenceToken

StartFromHead

StoredBytes

SubscriptionFilters

TagKey

TagList

TagValue

Tags

TargetArn

TestEventMessages

Timestamp

Token

Value