Crate rusoto_logs

You can use Amazon CloudWatch Logs to monitor, store, and access your log files from EC2 instances, Amazon CloudTrail, or other sources. You can then retrieve the associated log data from CloudWatch Logs using the Amazon CloudWatch console, the CloudWatch Logs commands in the AWS CLI, the CloudWatch Logs API, or the CloudWatch Logs SDK.

You can use CloudWatch Logs to:

Monitor Logs from Amazon EC2 Instances in Real-time: You can use CloudWatch Logs to monitor applications and systems using log data. For example, CloudWatch Logs can track the number of errors that occur in your application logs and send you a notification whenever the rate of errors exceeds a threshold you specify. CloudWatch Logs uses your log data for monitoring; so, no code changes are required. For example, you can monitor application logs for specific literal terms (such as "NullReferenceException") or count the number of occurrences of a literal term at a particular position in log data (such as "404" status codes in an Apache access log). When the term you are searching for is found, CloudWatch Logs reports the data to a Amazon CloudWatch metric that you specify.

Monitor Amazon CloudTrail Logged Events: You can create alarms in Amazon CloudWatch and receive notifications of particular API activity as captured by CloudTrail and use the notification to perform troubleshooting.

Archive Log Data: You can use CloudWatch Logs to store your log data in highly durable storage. You can change the log retention setting so that any log events older than this setting are automatically deleted. The CloudWatch Logs agent makes it easy to quickly send both rotated and non-rotated log data off of a host and into the log service. You can then access the raw log data when you need it.

If you're using the service, you're probably looking for CloudWatchLogsClient and CloudWatchLogs.

Structs

CancelExportTaskRequest

CloudWatchLogsClient
A client for the Amazon CloudWatch Logs API.

CreateExportTaskRequest

CreateExportTaskResponse

CreateLogGroupRequest

CreateLogStreamRequest

DeleteDestinationRequest

DeleteLogGroupRequest

DeleteLogStreamRequest

DeleteMetricFilterRequest

DeleteRetentionPolicyRequest

DeleteSubscriptionFilterRequest

DescribeDestinationsRequest

DescribeDestinationsResponse

DescribeExportTasksRequest

DescribeExportTasksResponse

DescribeLogGroupsRequest

DescribeLogGroupsResponse

DescribeLogStreamsRequest

DescribeLogStreamsResponse

DescribeMetricFiltersRequest

DescribeMetricFiltersResponse

DescribeSubscriptionFiltersRequest

DescribeSubscriptionFiltersResponse

Destination
Represents a cross-account destination that receives subscription log events.

ExportTask
Represents an export task.

ExportTaskExecutionInfo
Represents the status of an export task.

ExportTaskStatus
Represents the status of an export task.

FilterLogEventsRequest

FilterLogEventsResponse

FilteredLogEvent
Represents a matched event.

GetLogEventsRequest

GetLogEventsResponse

InputLogEvent
Represents a log event, which is a record of activity that was recorded by the application or resource being monitored.

ListTagsLogGroupRequest

ListTagsLogGroupResponse

LogGroup
Represents a log group.

LogStream
Represents a log stream, which is a sequence of log events from a single emitter of logs.

MetricFilter
Metric filters express how CloudWatch Logs would extract metric observations from ingested log events and transform them into metric data in a CloudWatch metric.

MetricFilterMatchRecord
Represents a matched event.

MetricTransformation
Indicates how to transform ingested log events into metric data in a CloudWatch metric.

OutputLogEvent
Represents a log event.

PutDestinationPolicyRequest

PutDestinationRequest

PutDestinationResponse

PutLogEventsRequest

PutLogEventsResponse

PutMetricFilterRequest

PutRetentionPolicyRequest

PutSubscriptionFilterRequest

RejectedLogEventsInfo
Represents the rejected events.

SearchedLogStream
Represents the search status of a log stream.

SubscriptionFilter
Represents a subscription filter.

TagLogGroupRequest

TestMetricFilterRequest

TestMetricFilterResponse

UntagLogGroupRequest

Enums

CancelExportTaskError
Errors returned by CancelExportTask

CreateExportTaskError
Errors returned by CreateExportTask

CreateLogGroupError
Errors returned by CreateLogGroup

CreateLogStreamError
Errors returned by CreateLogStream

DeleteDestinationError
Errors returned by DeleteDestination

DeleteLogGroupError
Errors returned by DeleteLogGroup

DeleteLogStreamError
Errors returned by DeleteLogStream

DeleteMetricFilterError
Errors returned by DeleteMetricFilter

DeleteRetentionPolicyError
Errors returned by DeleteRetentionPolicy

DeleteSubscriptionFilterError
Errors returned by DeleteSubscriptionFilter

DescribeDestinationsError
Errors returned by DescribeDestinations

DescribeExportTasksError
Errors returned by DescribeExportTasks

DescribeLogGroupsError
Errors returned by DescribeLogGroups

DescribeLogStreamsError
Errors returned by DescribeLogStreams

DescribeMetricFiltersError
Errors returned by DescribeMetricFilters

DescribeSubscriptionFiltersError
Errors returned by DescribeSubscriptionFilters

FilterLogEventsError
Errors returned by FilterLogEvents

GetLogEventsError
Errors returned by GetLogEvents

ListTagsLogGroupError
Errors returned by ListTagsLogGroup

PutDestinationError
Errors returned by PutDestination

PutDestinationPolicyError
Errors returned by PutDestinationPolicy

PutLogEventsError
Errors returned by PutLogEvents

PutMetricFilterError
Errors returned by PutMetricFilter

PutRetentionPolicyError
Errors returned by PutRetentionPolicy

PutSubscriptionFilterError
Errors returned by PutSubscriptionFilter

TagLogGroupError
Errors returned by TagLogGroup

TestMetricFilterError
Errors returned by TestMetricFilter

UntagLogGroupError
Errors returned by UntagLogGroup

Traits

CloudWatchLogs
Trait representing the capabilities of the Amazon CloudWatch Logs API. Amazon CloudWatch Logs clients implement this trait.